refactor(core): resolve circular dependencies
Eliminated circular dependencies in the core module and reorganized package structure.

Ref #2321
