+++ packages/core/src/i18n/i18nUtil.ts
@@ -1,8 +1,6 @@
 import type { ErrorObject } from 'ajv';
-import { isInternationalized, Labelable, UISchemaElement } from '../models';
-import { getControlPath } from '../reducers';
-import { formatErrorMessage } from '../util';
-import type { i18nJsonSchema, ErrorTranslator, Translator } from './i18nTypes';
+import { Labelable, UISchemaElement } from '../models';
+import type { i18nJsonSchema, ErrorTranslator, Translator } from '../store';
 import {
   ArrayDefaultTranslation,
   ArrayTranslations,
@@ -11,6 +9,11 @@ import {
   CombinatorDefaultTranslation,
   CombinatorTranslations,
 } from './combinatorTranslations';
+import {
+  formatErrorMessage,
+  getControlPath,
+  isInternationalized,
+} from '../util';
 
 export const getI18nKeyPrefixBySchema = (
   schema: i18nJsonSchema | undefined,
