+++ packages/core/src/i18n/index.ts
@@ -1,4 +1,4 @@
-export * from './i18nTypes';
-export * from './i18nUtil';
 export * from './arrayTranslations';
 export * from './combinatorTranslations';
+export * from './selectors';
+export * from './i18nUtil';
