+++ packages/core/src/mappers/index.ts
@@ -1,19 +1,19 @@
 /*
   The MIT License
-
-  Copyright (c) 2017-2019 EclipseSource Munich
+  
+  Copyright (c) 2024 EclipseSource Munich
   https://github.com/eclipsesource/jsonforms
-
+  
   Permission is hereby granted, free of charge, to any person obtaining a copy
   of this software and associated documentation files (the "Software"), to deal
   in the Software without restriction, including without limitation the rights
   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
   copies of the Software, and to permit persons to whom the Software is
   furnished to do so, subject to the following conditions:
-
+  
   The above copyright notice and this permission notice shall be included in
   all copies or substantial portions of the Software.
-
+  
   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
   IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,37 +23,7 @@
   THE SOFTWARE.
 */
 
-import find from 'lodash/find';
-import type { JsonSchema } from '../models';
-
-export const getFirstPrimitiveProp = (schema: any) => {
-  if (schema.properties) {
-    return find(Object.keys(schema.properties), (propName) => {
-      const prop = schema.properties[propName];
-      return (
-        prop.type === 'string' ||
-        prop.type === 'number' ||
-        prop.type === 'integer'
-      );
-    });
-  }
-  return undefined;
-};
-
-/**
- * Tests whether the schema has an enum based on oneOf.
- */
-export const isOneOfEnumSchema = (schema: JsonSchema) =>
-  !!schema &&
-  Object.prototype.hasOwnProperty.call(schema, 'oneOf') &&
-  schema.oneOf &&
-  (schema.oneOf as JsonSchema[]).every((s) => s.const !== undefined);
-
-/**
- * Tests whether the schema has an enum.
- */
-export const isEnumSchema = (schema: JsonSchema) =>
-  !!schema &&
-  typeof schema === 'object' &&
-  (Object.prototype.hasOwnProperty.call(schema, 'enum') ||
-    Object.prototype.hasOwnProperty.call(schema, 'const'));
+export * from './renderer';
+export * from './cell';
+export * from './combinators';
+export * from './util';
