+++ packages/core/src/models/uischema.ts
@@ -284,32 +284,3 @@ export interface Categorization
    */
   elements: (Category | Categorization)[];
 }
-
-export const isInternationalized = (
-  element: unknown
-): element is Required<Internationalizable> =>
-  typeof element === 'object' &&
-  element !== null &&
-  typeof (element as Internationalizable).i18n === 'string';
-
-export const isGroup = (layout: Layout): layout is GroupLayout =>
-  layout.type === 'Group';
-
-export const isLayout = (uischema: UISchemaElement): uischema is Layout =>
-  (uischema as Layout).elements !== undefined;
-
-export const isScopable = (obj: unknown): obj is Scopable =>
-  !!obj && typeof obj === 'object';
-
-export const isScoped = (obj: unknown): obj is Scoped =>
-  isScopable(obj) && typeof obj.scope === 'string';
-
-export const isLabelable = (obj: unknown): obj is Labelable =>
-  !!obj && typeof obj === 'object';
-
-export const isLabeled = <T = never>(obj: unknown): obj is Labeled<T> =>
-  isLabelable(obj) && ['string', 'boolean'].includes(typeof obj.label);
-
-export const isControlElement = (
-  uiSchema: UISchemaElement
-): uiSchema is ControlElement => uiSchema.type === 'Control';
