+++ packages/core/src/reducers/default-data.ts
@@ -23,13 +23,15 @@
   THE SOFTWARE.
 */
 
+import get from 'lodash/get';
 import {
   ADD_DEFAULT_DATA,
-  RegisterDefaultDataAction,
   REMOVE_DEFAULT_DATA,
+  RegisterDefaultDataAction,
   UnregisterDefaultDataAction,
 } from '../actions';
-import type { Reducer } from '../util';
+import type { Reducer } from '../store/type';
+import { JsonFormsState } from '../store';
 
 export interface JsonFormsDefaultDataRegistryEntry {
   schemaPath: string;
@@ -56,6 +58,10 @@ export const defaultDataReducer: Reducer<
   }
 };
 
+export const getDefaultData = (
+  state: JsonFormsState
+): JsonFormsDefaultDataRegistryEntry[] =>
+  extractDefaultData(get(state, 'jsonforms.defaultData'));
 export const extractDefaultData = (
   state: JsonFormsDefaultDataRegistryEntry[]
 ): JsonFormsDefaultDataRegistryEntry[] => state;
