+++ packages/core/src/store/index.ts
@@ -1,7 +1,7 @@
 /*
   The MIT License
 
-  Copyright (c) 2017-2019 EclipseSource Munich
+  Copyright (c) 2024-2024 EclipseSource Munich
   https://github.com/eclipsesource/jsonforms
 
   Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -23,21 +23,7 @@
   THE SOFTWARE.
 */
 
-const move = (array: any[], index: number, delta: number) => {
-  const newIndex: number = index + delta;
-  if (newIndex < 0 || newIndex >= array.length) {
-    return;
-  } // Already at the top or bottom.
-  const indexes: number[] = [index, newIndex].sort((a, b) => a - b); // Sort the indixes
-  array.splice(indexes[0], 2, array[indexes[1]], array[indexes[0]]);
-};
-
-const moveUp = (array: any[], toMove: number) => {
-  move(array, toMove, -1);
-};
-
-const moveDown = (array: any[], toMove: number) => {
-  move(array, toMove, 1);
-};
-
-export { moveUp, moveDown };
+export * from './store';
+export * from './type';
+export * from './i18nTypes';
+export * from './jsonFormsCore';
