+++ packages/core/src/util/index.ts
@@ -23,19 +23,16 @@
   THE SOFTWARE.
 */
 
-export * from './array';
-export * from './cell';
-export * from './combinators';
 export * from './Formatted';
 export * from './ids';
 export * from './label';
 export * from './path';
-export * from './renderer';
 export * from './resolvers';
 export * from './runtime';
 export * from './schema';
-export * from './type';
 export * from './uischema';
 export * from './util';
 export * from './validator';
 export * from './defaultDateFormat';
+export * from './errors';
+export * from './helpers';
