+++ packages/core/src/util/runtime.ts
@@ -27,7 +27,6 @@ import has from 'lodash/has';
 import {
   AndCondition,
   Condition,
-  JsonSchema,
   LeafCondition,
   OrCondition,
   RuleEffect,
@@ -36,10 +35,8 @@ import {
   UISchemaElement,
 } from '../models';
 import { resolveData } from './resolvers';
-import { composeWithUi } from './path';
 import type Ajv from 'ajv';
-import { getAjv } from '../reducers';
-import type { JsonFormsState } from '../store';
+import { composeWithUi } from './uischema';
 
 const isOrCondition = (condition: Condition): condition is OrCondition =>
   condition.type === 'OR';
@@ -182,45 +179,5 @@ export const isEnabled = (
     return evalEnablement(uischema, data, path, ajv);
   }
 
-  return true;
-};
-
-/**
- * Indicates whether the given `uischema` element shall be enabled or disabled.
- * Checks the global readonly flag, uischema rule, uischema options (including the config),
- * the schema and the enablement indicator of the parent.
- */
-export const isInherentlyEnabled = (
-  state: JsonFormsState,
-  ownProps: any,
-  uischema: UISchemaElement,
-  schema: (JsonSchema & { readOnly?: boolean }) | undefined,
-  rootData: any,
-  config: any
-) => {
-  if (state?.jsonforms?.readonly) {
-    return false;
-  }
-  if (uischema && hasEnableRule(uischema)) {
-    return isEnabled(uischema, rootData, ownProps?.path, getAjv(state));
-  }
-  if (typeof uischema?.options?.readonly === 'boolean') {
-    return !uischema.options.readonly;
-  }
-  if (typeof uischema?.options?.readOnly === 'boolean') {
-    return !uischema.options.readOnly;
-  }
-  if (typeof config?.readonly === 'boolean') {
-    return !config.readonly;
-  }
-  if (typeof config?.readOnly === 'boolean') {
-    return !config.readOnly;
-  }
-  if (schema?.readOnly === true) {
-    return false;
-  }
-  if (typeof ownProps?.enabled === 'boolean') {
-    return ownProps.enabled;
-  }
   return true;
 };
