+++ packages/core/src/util/validator.ts
@@ -24,7 +24,7 @@
 */
 import Ajv from 'ajv';
 import addFormats from 'ajv-formats';
-import type { Options } from 'ajv';
+import type { ErrorObject, Options, ValidateFunction } from 'ajv';
 
 export const createAjv = (options?: Options) => {
   const ajv = new Ajv({
@@ -36,4 +36,18 @@ export const createAjv = (options?: Options) => {
   });
   addFormats(ajv);
   return ajv;
+};
+
+export const validate = (
+  validator: ValidateFunction | undefined,
+  data: any
+): ErrorObject[] => {
+  if (validator === undefined) {
+    return [];
+  }
+  const valid = validator(data);
+  if (valid) {
+    return [];
+  }
+  return validator.errors;
 };
